MISRA Rules related to: 2, 8, 10, 11 and DIR 4 YT ID: DO-08, DO-11, DO-17, DO-19, DO-20

+++ hal/armv7r/zynqmp/zynqmp.c
@@ -14,9 +14,11 @@
  */
 
 #include "hal/cpu.h"
+#include "hal/hal.h"
 #include "hal/armv7r/armv7r.h"
 #include "hal/spinlock.h"
 #include "include/arch/armv7r/zynqmp/zynqmp.h"
+#include "zynqmp.h"
 
 #include "hal/armv7r/halsyspage.h"
 #include "zynqmp_regs.h" /*TBC - hal/armv7r/zynqmp_regs.h*/
@@ -164,11 +166,11 @@ int _zynqmp_setMIO(unsigned pin, u8 l0, u8 l1, u8 l2, u8 l3, u8 config)
 
 static void _zynqmp_getMIOMuxing(unsigned pin, u8 *l0, u8 *l1, u8 *l2, u8 *l3)
 {
-	u8 val = *(zynq_common.iou_slcr + iou_slcr_mio_pin_0 + pin) & 0xffU;
-	*l0 = (val >> 1) & 0x1U;
-	*l1 = (val >> 2) & 0x1U;
-	*l2 = (val >> 3) & 0x3U;
-	*l3 = (val >> 5) & 0x7U;
+	u32 val = *(zynq_common.iou_slcr + iou_slcr_mio_pin_0 + pin) & 0xffU;
+	*l0 = (u8)(val >> 1) & 0x1U;
+	*l1 = (u8)(val >> 2) & 0x1U;
+	*l2 = (u8)(val >> 3) & 0x3U;
+	*l3 = (u8)(val >> 5) & 0x7U;
 }
 
 
@@ -367,7 +369,7 @@ void hal_wdgReload(void)
 }
 
 
-int hal_platformctl(void *ptr)
+int hal_platformctl(void *ptr)  // TBD_Julia unused - delete or add static?
 {
 	platformctl_t *data = ptr;
 	unsigned int t = 0;
